Enable double precision for PyQG values.
This fixes our numerical issues for now at the cost of expensive
computation. I think the issues largely come from interactions with
invert(), particularly the result of _apply_a_ph which produces very
large values which are combined with our very small states, but there
may be other issues.

+++ src/systems/qg/kernel.py
@@ -14,8 +14,8 @@ def _generic_irfftn(a):
     return jnp.fft.irfftn(a, axes=(-2, -1))
 
 
-DTYPE_COMPLEX = jnp.complex64
-DTYPE_REAL = jnp.float32
+DTYPE_COMPLEX = jnp.complex128
+DTYPE_REAL = jnp.float64
 
 
 def register_dataclass_pytree(cls):
